# ProyectoFinalAvanzada/src/main/gen/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: API para reportar incidentes ciudadanos en tiempo real.
  title: Plataforma de Reportes de Seguridad
  version: 1.0.0
servers:
- description: Servidor de producción
  url: https://api.reportes-ciudad.com/v1
- description: Servidor local para desarrollo
  url: http://localhost:8080
paths:
...
  /notificaciones:
    get:
      description: Permite a un usuario ver todas sus notificaciones recibidas.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Notificacion'
                type: array
          description: Lista de notificaciones obtenida con éxito.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_notificaciones_get_401_response'
          description: No autorizado.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_notificaciones_get_500_response'
          description: Error interno del servidor.
      summary: Obtener notificaciones del usuario
      tags:
      - notifications
      x-accepts:
      - application/json
      x-tags:
      - tag: notifications
    post:
      description: Envía una notificación a usuarios que se encuentran cerca del reporte.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notificacion'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_notificaciones_post_200_response'
          description: Notificación enviada con éxito.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_notificaciones_post_400_response'
          description: Datos inválidos.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_notificaciones_post_500_response'
          description: Error interno del servidor.
      summary: Enviar notificación a usuarios cercanos
      tags:
      - notifications
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: notifications
  /notificaciones/{id}/leer:
    put:
      description: Permite a un usuario marcar una notificación como leída.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Notificación marcada como leída.
        "404":
          description: Notificación no encontrada.
      summary: Marcar notificación como leída
      tags:
      - notifications
      x-accepts:
      - application/json
      x-tags:
      - tag: notifications
  /notificaciones/{id}:
    get:
      description: Devuelve la información completa de una notificación específica.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notificacion'
          description: Notificación encontrada.
        "404":
          description: Notificación no encontrada.
      summary: Obtener detalles de una notificación
      tags:
      - notifications
      x-accepts:
      - application/json
      x-tags:
      - tag: notifications
components:
  schemas:
...
    Notificacion:
      example:
        fecha: 2024-03-10T14:30:00Z
        mensaje: Nuevo reporte de accidente cercano a tu ubicación.
        enlaceReporte: https://miapp.com/reportes/reporte567
        reporteId: reporte567
        usuarioId: abc123
      properties:
        usuarioId:
          example: abc123
          type: string
        reporteId:
          example: reporte567
          type: string
        mensaje:
          example: Nuevo reporte de accidente cercano a tu ubicación.
          type: string
        fecha:
          example: 2024-03-10T14:30:00Z
          format: date-time
          type: string
        enlaceReporte:
          example: https://miapp.com/reportes/reporte567
          format: uri
          type: string
      required:
      - fecha
      - mensaje
      - reporteId
      - usuarioId
      type: object
...
    _notificaciones_get_401_response:
      example:
        error: Debes iniciar sesión para ver tus notificaciones.
      properties:
        error:
          example: Debes iniciar sesión para ver tus notificaciones.
          type: string
      type: object
    _notificaciones_get_500_response:
      example:
        error: Error inesperado al obtener las notificaciones.
      properties:
        error:
          example: Error inesperado al obtener las notificaciones.
          type: string
      type: object
    _notificaciones_post_200_response:
      example:
        mensaje: Notificaciones enviadas correctamente.
      properties:
        mensaje:
          example: Notificaciones enviadas correctamente.
          type: string
      type: object
    _notificaciones_post_400_response:
      example:
        error: Faltan datos en la notificación.
      properties:
        error:
          example: Faltan datos en la notificación.
          type: string
      type: object
    _notificaciones_post_500_response:
      example:
        error: Error inesperado al enviar las notificaciones.
      properties:
        error:
          example: Error inesperado al enviar las notificaciones.
          type: string
      type: object
...